Add TODOs and helper code to skeleton pass
We're using the skeleton pass for the first set of tasks in the workshop. These
deal with both IR introspection and a bit of transformation. The end goal is to
find all calls to the 'fib' function with a constant parameter, and replace them
with the precomputed result.

This commit adds progressive TODOs to that end, as well as a couple of helper
methods for identifying 'fib' calls (checking the name and prototype of the
called function) and for getting the precomputed result of a call to 'fib'.
Buyer beware, the helpers are a bit sloppy.