Add code skeleton for static callgraph pass
Add the code skeleton for the workshop tasks that deal with obtaining the static
callgraph of programs. Also add some targets in the sample directory to run the
static callgraph pass on the 'd.c' sample.

The pass uses a utility library called 'callgraph' for dumping the callgraph.
The library will also be used as a runtime library for a future task (dynamic
callgraph). It must be installed on the system before these passes can work. We
should probably integrate it into CMake properly, instead of requiring it to be
installed.